U and D of CRUD backend

+++ routes/api/questions.js
@@ -3,6 +3,7 @@ const router = express.Router();
 const passport = require("passport");
 const validateQuestionInput = require('../../validation/question');
 const Question = require('../../models/Question');
+const User = require('../../models/User');
 
 
 //test
@@ -29,23 +30,32 @@ router.get('/:id',(req,res)=>{
 })
 
 //posting questions
-router.post('/',
-    passport.authenticate('jwt',{session:false}),
-    (req,res) =>{
+router.post('/',passport.authenticate('jwt',{session:false}),(req,res) =>{
         //check validation
         const {errors, isValid} = validateQuestionInput(req.body);
-        if(!isValid){
-            return res.status(400).json(errors);
-        }
-        const newQuestion = new Question({
-            subject: req.body.subject,
-            content: req.body.content,
-            tag: req.body.tag,
-            solved: req.body.solved
-
-    });
-
-    newQuestion.save().then(question => res.json(question));
+         if (!isValid) {
+             return res.status(400).json(errors);
+         }
+        Question.findOne({subject:req.body.subject}).then( question => {
+            if (question) {
+                return res.status(400).json({
+                Error: "Question has already been submitted"
+                })       
+            } 
+            else {
+              
+               const newQuestion = new Question({
+                user: req.user.id,
+                subject: req.body.subject,
+                content: req.body.content,
+                tag: req.body.tag,
+                solved: req.body.solved
+        });
+         newQuestion.save().then(question => res.json(question));
+      }
+    
+        })
+    
 });
 
 
@@ -61,44 +71,63 @@ router.post('/',
 
 
 
-router.patch("/:id", async (req, res) => {
-	try {
-		const question = await Question.findById(req.params._id)
-
-		if (req.body.subject) {
-			question.subject = req.body.subject
-		}
-
-		if (req.body.content) {
-			question.content = req.body.content
-        }
+router.patch("/:id", passport.authenticate('jwt',{session:false}), async (req, res) => {
+    try {
+        let question = await Question.findById(req.params.id)
+        // console.log("user: " + question.user)
+        // console.log("req user " + req.user.id )
+        if(req.user.id ===  `${question.user}`){
         
-        if (req.body.tag) {
-            question.tag = req.body.tag
+            if (req.body.content) {
+                question.content = req.body.content
+            }
+            
+            if (req.body.tag) {
+                question.tag = req.body.tag
+            }
+
+            if (req.body.solved) {
+                question.solved = req.body.solved
+            }
+
+            await question.save()
+            res.send(question)
+        }else {
+        res.status(404).json({
+            error: 'Incorrect user'
+        })
+
         }
-
-          if (req.body.solved) {
-              question.solved = req.body.solved
-          }
-
-		await question.save()
-		res.send(question)
-	} catch {
-		res.status(404)
-		res.send({ error: "Question doesn't exist!" })
-	}
+	} catch(err) {
+        res.json(err)
+        res.status(404).json({
+            error: "Question doesn't exist!"
+        })
+    }
+    
 })
 
 
 
-//! CURRENTLY ONLY UPDATES CONTENT
 
 //deleting a question
-router.delete("/:id", (req, res) => {
-  Question.findByIdAndDelete(req.params.id)
-    .then(() => res.json("Question deleted"))
-    .catch(err => res.status(404).json(err))
+router.delete("/:id", passport.authenticate('jwt',{session:false}), async (req, res) => {
+
+    // console.log('1', req.user.id)
+    const question = await Question.findOne({ _id: req.params.id })
+    // console.log('2', question.user)
+
+    if (`${question.user}` === req.user.id){
+        Question.findByIdAndDelete(req.params.id)
+        .then(() => res.json("Question deleted"))
+        .catch(err => res.status(404).json(err))
+    } else{
+        res.status(404).json({error: 'Incorrect user'})
+    }
+    
 })
-//!TEST FOR UNIQUENESS OF SUBJECT
 
 module.exports = router;
+
+
+
